actually only run requests once
